requirements.txt
alembic==0.9.7
aniso8601==2.0.0
asn1crypto==0.24.0
cffi==1.11.4
click==6.7
cryptography==2.1.4
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jwcrypto==0.4.2
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python==2.1.7
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
six==1.11.0
SQLAlchemy==1.2.2
Werkzeug==0.14.1